setting proper contact values for shin and foot
Hi all I will summarize the results of our difficult and painful
debugging month

1) in  Gazebo 2.0

the properties of a fake link (e.g. the link_foot) were inherithing the
properties of the parent foot (e.g. the lowerleg)

in our case to have a proper contact behaviour we were setting the
contact stiffness to kp = 1000000 and the damping to kd= 100

in Gazebo 7.0 this does not happen anymore, so the stiffness and the
damping for the foot were not kp = 1000000 and the damping to kd= 100

anymore but were set to the MAXIMUM possible value screwing the
simulation (the robot glides...)

2) to have meaningful joint velocity readings we should set the
parameter <max_vel>1.0</max_vel>

otherwise it spits out crappy value (again this value was not set for
the foot and the joint velocities  were not meaningful)